requirements.txt
fastapi
uvicorn
python-multipart
pdf2image
Pillow
python-dotenv
google-generativeai
